refactor: moves enum trait to traits directory
