+++ tests/unit/Files/Enums/FileTypeEnumTest.php
@@ -6,7 +6,7 @@ namespace WordPress\AiClient\Tests\unit\Files\Enums;
 
 use PHPUnit\Framework\TestCase;
 use WordPress\AiClient\Files\Enums\FileTypeEnum;
-use WordPress\AiClient\Tests\unit\EnumTestTrait;
+use WordPress\AiClient\Tests\traits\EnumTestTrait;
 
 /**
  * @covers \WordPress\AiClient\Files\Enums\FileTypeEnum
