+++ tests/unit/Providers/Enums/ToolTypeEnumTest.php
@@ -6,7 +6,7 @@ namespace WordPress\AiClient\Tests\unit\Providers\Enums;
 
 use PHPUnit\Framework\TestCase;
 use WordPress\AiClient\Providers\Enums\ToolTypeEnum;
-use WordPress\AiClient\Tests\unit\EnumTestTrait;
+use WordPress\AiClient\Tests\traits\EnumTestTrait;
 
 /**
  * @covers \WordPress\AiClient\Providers\Enums\ToolTypeEnum
